chore: 유효성 체크 로직 분리
 - InputView 에서 입력값 유효성을 체크하는 로직 분리 (Validator.java)
